+++ extensions/summarize-youtube-video-with-ai/src/components/openai/hooks/useOpenAIFollowUpQuestion.ts
@@ -1,58 +1,90 @@
 import { getPreferenceValues, showToast, Toast } from "@raycast/api";
 import OpenAI from "openai";
+import { useEffect } from "react";
 import { OPENAI_MODEL } from "../../../const/defaults";
 import { ALERT, FINDING_ANSWER } from "../../../const/toast_messages";
+import { Question } from "../../../hooks/useQuestions";
 import { OpenAIPreferences } from "../../../summarizeVideoWithOpenAI";
+import { generateQuestionId } from "../../../utils/generateQuestionId";
 import { getFollowUpQuestionSnippet } from "../../../utils/getAiInstructionSnippets";
 
-export const useOpenAIFollowUpQuestion = async (
-  question: string,
-  transcript: string,
-  setSummary: React.Dispatch<React.SetStateAction<string | undefined>>,
-  pop: () => void,
-) => {
-  const preferences = getPreferenceValues() as OpenAIPreferences;
-  const { openaiApiToken, openaiEndpoint, openaiModel } = preferences;
-  setSummary(undefined);
-
-  const openai = new OpenAI({
-    apiKey: openaiApiToken,
-  });
-
-  if (openaiEndpoint !== "") {
-    openai.baseURL = openaiEndpoint;
-  }
-
-  const toast = showToast({
-    style: Toast.Style.Animated,
-    title: FINDING_ANSWER.title,
-    message: FINDING_ANSWER.message,
-  });
-
-  const answer = openai.beta.chat.completions.stream({
-    model: openaiModel || OPENAI_MODEL,
-    messages: [{ role: "user", content: getFollowUpQuestionSnippet(question, transcript) }],
-    stream: true,
-  });
-
-  pop();
-
-  answer.on("content", (delta) => {
-    setSummary((result) => {
-      if (result === undefined) return delta;
-      return result + delta;
-    });
-  });
-
-  answer.finalChatCompletion().then(() => {
-    toast.then((t) => t.hide());
-  });
-
-  answer.on("error", (error) => {
-    toast.then((t) => {
-      t.style = Toast.Style.Failure;
-      t.title = ALERT.title;
-      t.message = error.message;
-    });
-  });
+type FollowUpQuestionParams = {
+  setQuestions: React.Dispatch<React.SetStateAction<Question[]>>;
+  setQuestion: React.Dispatch<React.SetStateAction<string>>;
+  transcript: string | undefined;
+  question: string;
 };
+
+export function useOpenAIFollowUpQuestion({ setQuestions, setQuestion, transcript, question }: FollowUpQuestionParams) {
+  const preferences = getPreferenceValues() as OpenAIPreferences;
+  const { openaiApiToken, openaiEndpoint, openaiModel, creativity } = preferences;
+
+  useEffect(() => {
+    const abortController = new AbortController();
+
+    const handleAdditionalQuestion = async () => {
+      if (!question || !transcript) return;
+      const qID = generateQuestionId();
+
+      const toast = await showToast({
+        style: Toast.Style.Animated,
+        title: FINDING_ANSWER.title,
+        message: FINDING_ANSWER.message,
+      });
+
+      const openai = new OpenAI({
+        apiKey: openaiApiToken,
+      });
+
+      if (openaiEndpoint !== "") {
+        openai.baseURL = openaiEndpoint;
+      }
+
+      setQuestions((prevQuestions) => [
+        {
+          id: qID,
+          question,
+          answer: "",
+        },
+        ...prevQuestions,
+      ]);
+
+      const answer = openai.beta.chat.completions.stream(
+        {
+          model: openaiModel || OPENAI_MODEL,
+          messages: [{ role: "user", content: getFollowUpQuestionSnippet(question, transcript) }],
+          stream: true,
+          creativity: parseInt(creativity),
+        },
+        { signal: abortController.signal },
+      );
+
+      answer.on("content", (delta) => {
+        toast.show();
+        setQuestions((prevQuestions) =>
+          prevQuestions.map((q) => (q.id === qID ? { ...q, answer: q.answer + delta } : q)),
+        );
+      });
+
+      answer.finalChatCompletion().then(() => {
+        toast.hide();
+        setQuestion("");
+      });
+
+      if (abortController.signal.aborted) return;
+
+      answer.on("error", (error) => {
+        if (abortController.signal.aborted) return;
+        toast.style = Toast.Style.Failure;
+        toast.title = ALERT.title;
+        toast.message = error.message;
+      });
+    };
+
+    handleAdditionalQuestion();
+
+    return () => {
+      abortController.abort();
+    };
+  }, [question, transcript]);
+}